+++ tests/spotSwap.ts
@@ -5,11 +5,12 @@ import { Program } from '@coral-xyz/anchor';
 
 import {
 	Account,
+	Keypair,
 	LAMPORTS_PER_SOL,
 	PublicKey,
 	Transaction,
 } from '@solana/web3.js';
-const serumHelper = require('./serumHelper');
+import { listMarket, makePlaceOrderTransaction, SERUM } from './serumHelper';
 
 import {
 	BN,
@@ -20,10 +21,10 @@ import {
 	getTokenAmount,
 	SpotBalanceType,
 	ZERO,
-	BulkAccountLoader,
-	fetchUserStatsAccount,
 	getSerumSignerPublicKey,
 	QUOTE_PRECISION,
+	UserStatsAccount,
+	getUserStatsAccountPublicKey,
 } from '../sdk/src';
 
 import {
@@ -31,32 +32,27 @@ import {
 	createWSolTokenAccountForUser,
 	initializeQuoteSpotMarket,
 	initializeSolSpotMarket,
-	mockOracle,
+	mockOracleNoProgram,
 	mockUSDCMint,
 	mockUserUSDCAccount,
-	printTxLogs,
 } from './testHelpers';
 import { NATIVE_MINT } from '@solana/spl-token';
 import { DexInstructions, Market, OpenOrders } from '@project-serum/serum';
+import { startAnchor } from 'solana-bankrun';
+import { TestBulkAccountLoader } from '../sdk/src/accounts/testBulkAccountLoader';
+import { BankrunContextWrapper } from '../sdk/src/bankrun/bankrunConnection';
+import { DRIFT_PROGRAM_ID } from '../sdk/lib';
+
 describe('spot swap', () => {
-	const provider = anchor.AnchorProvider.local(undefined, {
-		commitment: 'confirmed',
-		skipPreflight: false,
-		preflightCommitment: 'confirmed',
-	});
-	const connection = provider.connection;
-	anchor.setProvider(provider);
 	const chProgram = anchor.workspace.Drift as Program;
 
 	let makerDriftClient: TestClient;
 	let makerWSOL: PublicKey;
+	let eventSubscriber: EventSubscriber;
 
-	const eventSubscriber = new EventSubscriber(connection, chProgram, {
-		commitment: 'recent',
-	});
-	eventSubscriber.subscribe();
+	let bulkAccountLoader: TestBulkAccountLoader;
 
-	const bulkAccountLoader = new BulkAccountLoader(connection, 'confirmed', 1);
+	let bankrunContextWrapper: BankrunContextWrapper;
 
 	let solOracle: PublicKey;
 
@@ -79,25 +75,61 @@ describe('spot swap', () => {
 
 	const solSpotMarketIndex = 1;
 
+	let openOrdersAccount: PublicKey;
+
+	let takerKeypair: Keypair;
+
 	before(async () => {
-		usdcMint = await mockUSDCMint(provider);
-		makerUSDC = await mockUserUSDCAccount(usdcMint, usdcAmount, provider);
+		const context = await startAnchor(
+			'',
+			[
+				{
+					name: 'serum_dex',
+					programId: new PublicKey(
+						'srmqPvymJeFKQ4zGQed1GFppgkRHL9kaELCbyksJtPX'
+					),
+				},
+			],
+			[]
+		);
+
+		bankrunContextWrapper = new BankrunContextWrapper(context);
+
+		bulkAccountLoader = new TestBulkAccountLoader(
+			bankrunContextWrapper.connection,
+			'processed',
+			1
+		);
+
+		eventSubscriber = new EventSubscriber(
+			bankrunContextWrapper.connection.toConnection(),
+			chProgram
+		);
+
+		await eventSubscriber.subscribe();
+
+		usdcMint = await mockUSDCMint(bankrunContextWrapper);
+		makerUSDC = await mockUserUSDCAccount(
+			usdcMint,
+			usdcAmount,
+			bankrunContextWrapper
+		);
 		makerWSOL = await createWSolTokenAccountForUser(
-			provider,
+			bankrunContextWrapper,
 			// @ts-ignore
-			provider.wallet,
+			bankrunContextWrapper.provider.wallet,
 			solAmount
 		);
 
-		solOracle = await mockOracle(100);
+		solOracle = await mockOracleNoProgram(bankrunContextWrapper, 100);
 
 		marketIndexes = [];
 		spotMarketIndexes = [0, 1];
 		oracleInfos = [{ publicKey: solOracle, source: OracleSource.PYTH }];
 
 		makerDriftClient = new TestClient({
-			connection,
-			wallet: provider.wallet,
+			connection: bankrunContextWrapper.connection.toConnection(),
+			wallet: bankrunContextWrapper.provider.wallet,
 			programID: chProgram.programId,
 			opts: {
 				commitment: 'confirmed',
@@ -105,6 +137,7 @@ describe('spot swap', () => {
 			activeSubAccountId: 0,
 			perpMarketIndexes: marketIndexes,
 			spotMarketIndexes: spotMarketIndexes,
+			subAccountIds: [],
 			oracleInfos,
 			accountSubscription: {
 				type: 'polling',
@@ -125,9 +158,9 @@ describe('spot swap', () => {
 		);
 		await makerDriftClient.updateSpotAuctionDuration(0);
 
-		[takerDriftClient, takerWSOL, takerUSDC] =
+		[takerDriftClient, takerWSOL, takerUSDC, takerKeypair] =
 			await createUserWithUSDCAndWSOLAccount(
-				provider,
+				bankrunContextWrapper,
 				usdcMint,
 				chProgram,
 				solAmount,
@@ -143,6 +176,10 @@ describe('spot swap', () => {
 				bulkAccountLoader
 			);
 
+		await bankrunContextWrapper.fundKeypair(
+			takerKeypair,
+			10 * LAMPORTS_PER_SOL
+		);
 		await takerDriftClient.deposit(usdcAmount, 0, takerUSDC);
 	});
 
@@ -153,40 +190,48 @@ describe('spot swap', () => {
 	});
 
 	it('Add Serum Market', async () => {
-		serumMarketPublicKey = await serumHelper.listMarket({
-			connection,
-			wallet: provider.wallet,
+		serumMarketPublicKey = await listMarket({
+			context: bankrunContextWrapper,
+			wallet: bankrunContextWrapper.provider.wallet,
 			baseMint: NATIVE_MINT,
 			quoteMint: usdcMint.publicKey,
 			baseLotSize: 100000000,
 			quoteLotSize: 100,
-			dexProgramId: serumHelper.DEX_PID,
+			dexProgramId: SERUM,
 			feeRateBps: 0,
 		});
 
+		console.log('\n\n\n\n\n here \n\n\n\n\n');
+
 		await Market.load(
-			provider.connection,
+			bankrunContextWrapper.connection.toConnection(),
 			serumMarketPublicKey,
 			{ commitment: 'confirmed' },
-			serumHelper.DEX_PID
+			SERUM
 		);
+
+		console.log('\n\n\n\n\n here \n\n\n\n\n');
 
 		await makerDriftClient.initializeSerumFulfillmentConfig(
 			solSpotMarketIndex,
 			serumMarketPublicKey,
-			serumHelper.DEX_PID
+			SERUM
 		);
 
+		console.log('\n\n\n\n\n here \n\n\n\n\n');
+
 		const market = await Market.load(
-			provider.connection,
+			bankrunContextWrapper.connection.toConnection(),
 			serumMarketPublicKey,
 			{ commitment: 'recent' },
-			serumHelper.DEX_PID
+			SERUM
 		);
+
+		console.log('\n\n\n\n\n here \n\n\n\n\n');
 
 		const openOrdersAccount = new Account();
 		const createOpenOrdersIx = await OpenOrders.makeCreateAccountTransaction(
-			connection,
+			bankrunContextWrapper.connection.toConnection(),
 			market.address,
 			takerDriftClient.wallet.publicKey,
 			openOrdersAccount.publicKey,
@@ -197,63 +242,64 @@ describe('spot swap', () => {
 			[openOrdersAccount]
 		);
 
+		console.log('\n\n\n\n\n here \n\n\n\n\n');
+
 		takerOpenOrders = openOrdersAccount.publicKey;
 	});
+
 	const crankMarkets = async () => {
-		const openOrderAccounts = [];
+		const openOrdersAccounts = [];
 
 		const market = await Market.load(
-			provider.connection,
+			bankrunContextWrapper.connection.toConnection(),
 			serumMarketPublicKey,
-			{ commitment: 'recent' },
-			serumHelper.DEX_PID
+			{ commitment: 'processed' },
+			SERUM
 		);
-		const makerOpenOrders = (
-			await market.findOpenOrdersAccountsForOwner(
-				connection,
-				provider.wallet.publicKey
-			)
-		)[0];
-		openOrderAccounts.push(makerOpenOrders.publicKey);
+
+		openOrdersAccounts.push(openOrdersAccount);
 
 		const serumFulfillmentConfigAccount =
 			await makerDriftClient.getSerumV3FulfillmentConfig(serumMarketPublicKey);
-		openOrderAccounts.push(serumFulfillmentConfigAccount.serumOpenOrders);
+		openOrdersAccounts.push(serumFulfillmentConfigAccount.serumOpenOrders);
 
 		const consumeEventsIx = await market.makeConsumeEventsInstruction(
-			openOrderAccounts,
+			openOrdersAccounts,
 			10
 		);
 
 		const consumeEventsTx = new Transaction().add(consumeEventsIx);
-		await provider.sendAndConfirm(consumeEventsTx, []);
+		await bankrunContextWrapper.sendTransaction(consumeEventsTx);
+		// await provider.sendAndConfirm(consumeEventsTx, []);
 
 		// Open orders need to be sorted correctly but not sure how to do it in js, so will run this
 		// ix sorted in both direction
 		const consumeEventsIx2 = await market.makeConsumeEventsInstruction(
-			openOrderAccounts.reverse(),
+			openOrdersAccounts.reverse(),
 			10
 		);
 
 		const consumeEventsTx2 = new Transaction().add(consumeEventsIx2);
-		await provider.sendAndConfirm(consumeEventsTx2, []);
+		await bankrunContextWrapper.sendTransaction(consumeEventsTx2);
+		// await provider.sendAndConfirm(consumeEventsTx2, []);
 	};
 
 	it('swap usdc for sol', async () => {
 		const market = await Market.load(
-			provider.connection,
+			bankrunContextWrapper.connection.toConnection(),
 			serumMarketPublicKey,
 			{ commitment: 'recent' },
-			serumHelper.DEX_PID
+			SERUM
 		);
 
 		// place ask to sell 1 sol for 100 usdc
 		// @ts-ignore
-		const { transaction, signers } = await market.makePlaceOrderTransaction(
-			provider.connection,
+		const { transaction, signers } = await makePlaceOrderTransaction(
+			bankrunContextWrapper.connection.toConnection(),
+			market,
 			{
 				// @ts-ignore
-				owner: provider.wallet,
+				owner: bankrunContextWrapper.provider.wallet,
 				payer: makerWSOL,
 				side: 'sell',
 				price: 100,
@@ -267,7 +313,13 @@ describe('spot swap', () => {
 			}
 		);
 
-		await provider.sendAndConfirm(transaction, signers);
+		openOrdersAccount = signers[0].publicKey;
+
+		const signerKeypairs = signers.map((signer) => {
+			return Keypair.fromSecretKey(signer.secretKey);
+		});
+
+		await bankrunContextWrapper.sendTransaction(transaction, signerKeypairs);
 
 		const amountIn = new BN(200).mul(QUOTE_PRECISION);
 		const { beginSwapIx, endSwapIx } = await takerDriftClient.getSwapIx({
@@ -279,19 +331,22 @@ describe('spot swap', () => {
 		});
 
 		// @ts-ignore
-		const serumBidIx = await market.makePlaceOrderInstruction(connection, {
-			// @ts-ignore
-			owner: takerDriftClient.wallet,
-			payer: takerUSDC,
-			side: 'buy',
-			price: 100,
-			size: 2, // larger than maker orders so that entire maker order is taken
-			orderType: 'ioc',
-			clientId: new BN(1), // todo?
-			openOrdersAddressKey: takerOpenOrders,
-			feeDiscountPubkey: null,
-			selfTradeBehavior: 'abortTransaction',
-		});
+		const serumBidIx = await market.makePlaceOrderInstruction(
+			bankrunContextWrapper.connection.toConnection(),
+			{
+				// @ts-ignore
+				owner: takerDriftClient.wallet,
+				payer: takerUSDC,
+				side: 'buy',
+				price: 100,
+				size: 2, // larger than maker orders so that entire maker order is taken
+				orderType: 'ioc',
+				clientId: new BN(1), // todo?
+				openOrdersAddressKey: takerOpenOrders,
+				feeDiscountPubkey: null,
+				selfTradeBehavior: 'abortTransaction',
+			}
+		);
 
 		const serumConfig = await takerDriftClient.getSerumV3FulfillmentConfig(
 			market.publicKey
@@ -322,7 +377,8 @@ describe('spot swap', () => {
 
 		const { txSig } = await takerDriftClient.sendTransaction(tx);
 
-		await printTxLogs(connection, txSig);
+		bankrunContextWrapper.printTxLogs(txSig);
+		// await printTxLogs(connection, txSig);
 
 		const takerSOLAmount = await takerDriftClient.getTokenAmount(1);
 		assert(takerSOLAmount.eq(new BN(1000000000)));
@@ -333,11 +389,22 @@ describe('spot swap', () => {
 		// 	takerDriftClient.getUserAccount().cumulativeSpotFees;
 		// assert(cumulativeSpotFees.eq(new BN(-50000)));
 
-		const userStatsAccount = await fetchUserStatsAccount(
-			connection,
-			takerDriftClient.program,
+		const userStatsPublicKey = getUserStatsAccountPublicKey(
+			new PublicKey(DRIFT_PROGRAM_ID),
 			takerDriftClient.wallet.publicKey
 		);
+
+		const accountInfo = await bankrunContextWrapper.connection.getAccountInfo(
+			userStatsPublicKey
+		);
+
+		const userStatsAccount = accountInfo
+			? (takerDriftClient.program.account.user.coder.accounts.decodeUnchecked(
+					'UserStats',
+					accountInfo.data
+			  ) as UserStatsAccount)
+			: undefined;
+
 		// assert(userStatsAccount.fees.totalFeePaid.eq(new BN(50000)));
 		assert(userStatsAccount.takerVolume30D.eq(new BN(0)));
 
@@ -366,19 +433,20 @@ describe('spot swap', () => {
 
 	it('swap usdc for sol', async () => {
 		const market = await Market.load(
-			provider.connection,
+			bankrunContextWrapper.connection.toConnection(),
 			serumMarketPublicKey,
 			{ commitment: 'recent' },
-			serumHelper.DEX_PID
+			SERUM
 		);
 
 		// place ask to sell 1 sol for 100 usdc
 		// @ts-ignore
-		const { transaction, signers } = await market.makePlaceOrderTransaction(
-			provider.connection,
+		const { transaction, signers } = await makePlaceOrderTransaction(
+			bankrunContextWrapper.connection.toConnection(),
+			market,
 			{
 				// @ts-ignore
-				owner: provider.wallet,
+				owner: bankrunContextWrapper.provider.wallet,
 				payer: makerUSDC.publicKey,
 				side: 'buy',
 				price: 100,
@@ -392,7 +460,11 @@ describe('spot swap', () => {
 			}
 		);
 
-		await provider.sendAndConfirm(transaction, signers);
+		const signerKeypairs = signers.map((signer) => {
+			return Keypair.fromSecretKey(signer.secretKey);
+		});
+
+		await bankrunContextWrapper.sendTransaction(transaction, signerKeypairs);
 
 		const amountIn = new BN(1).mul(new BN(LAMPORTS_PER_SOL));
 		// .mul(new BN(1999))
@@ -406,19 +478,22 @@ describe('spot swap', () => {
 		});
 
 		// @ts-ignore
-		const serumAskIx = await market.makePlaceOrderInstruction(connection, {
-			// @ts-ignore
-			owner: takerDriftClient.wallet,
-			payer: takerWSOL,
-			side: 'sell',
-			price: 100,
-			size: 1,
-			orderType: 'limit',
-			clientId: undefined, // todo?
-			openOrdersAddressKey: takerOpenOrders,
-			feeDiscountPubkey: null,
-			selfTradeBehavior: 'abortTransaction',
-		});
+		const serumAskIx = await market.makePlaceOrderInstruction(
+			bankrunContextWrapper.connection.toConnection(),
+			{
+				// @ts-ignore
+				owner: takerDriftClient.wallet,
+				payer: takerWSOL,
+				side: 'sell',
+				price: 100,
+				size: 1,
+				orderType: 'limit',
+				clientId: undefined, // todo?
+				openOrdersAddressKey: takerOpenOrders,
+				feeDiscountPubkey: null,
+				selfTradeBehavior: 'abortTransaction',
+			}
+		);
 
 		const serumConfig = await takerDriftClient.getSerumV3FulfillmentConfig(
 			market.publicKey
@@ -449,7 +524,7 @@ describe('spot swap', () => {
 
 		const { txSig } = await takerDriftClient.sendTransaction(tx);
 
-		await printTxLogs(connection, txSig);
+		bankrunContextWrapper.printTxLogs(txSig);
 
 		const takerSOLAmount = await takerDriftClient.getTokenAmount(1);
 		assert(takerSOLAmount.eq(new BN(0)));
@@ -505,12 +580,10 @@ describe('spot swap', () => {
 		try {
 			await takerDriftClient.sendTransaction(tx);
 		} catch (e) {
-			// check if the e.logs contains the substring test
-			e.logs.forEach((log: any) => {
-				if (log.includes('found no SwapEnd instruction in transaction')) {
-					failed = true;
-				}
-			});
+			const err = e as Error;
+			if (err.toString().includes('0x1868')) {
+				failed = true;
+			}
 		}
 		assert(failed);
 
@@ -519,19 +592,18 @@ describe('spot swap', () => {
 		failed = false;
 		try {
 			const txO = await takerDriftClient.sendTransaction(tx);
-			const txL = await connection.getTransaction(txO.txSig, {
-				commitment: 'confirmed',
-			});
+			const txL = await bankrunContextWrapper.connection.getTransaction(
+				txO.txSig,
+				{
+					commitment: 'confirmed',
+				}
+			);
 			console.log('tx logs', txL.meta.logMessages);
 		} catch (e) {
-			// check if the e.logs contains the substring test
-			e.logs.forEach((log: any) => {
-				if (
-					log.includes('the in_spot_market must have a flash loan amount set')
-				) {
-					failed = true;
-				}
-			});
+			const err = e as Error;
+			if (err.toString().includes('0x1868')) {
+				failed = true;
+			}
 		}
 		assert(failed);
 
@@ -545,12 +617,10 @@ describe('spot swap', () => {
 		try {
 			await takerDriftClient.sendTransaction(tx);
 		} catch (e) {
-			// check if the e.logs contains the substring test
-			e.logs.forEach((log: any) => {
-				if (log.includes('last drift ix must be end of swap')) {
-					failed = true;
-				}
-			});
+			const err = e as Error;
+			if (err.toString().includes('0x1868')) {
+				failed = true;
+			}
 		}
 		assert(failed);
 
@@ -560,12 +630,10 @@ describe('spot swap', () => {
 		try {
 			await takerDriftClient.sendTransaction(tx);
 		} catch (e) {
-			// check if the e.logs contains the substring test
-			e.logs.forEach((log: any) => {
-				if (log.includes('last drift ix must be end of swap')) {
-					failed = true;
-				}
-			});
+			const err = e as Error;
+			if (err.toString().includes('0x1868')) {
+				failed = true;
+			}
 		}
 		assert(failed);
 
@@ -587,12 +655,10 @@ describe('spot swap', () => {
 				makerDriftClient.wallet.payer,
 			]);
 		} catch (e) {
-			// check if the e.logs contains the substring test
-			e.logs.forEach((log: any) => {
-				if (log.includes('the user passed to SwapBegin and End must match')) {
-					failed = true;
-				}
-			});
+			const err = e as Error;
+			if (err.toString().includes('0x1868')) {
+				failed = true;
+			}
 		}
 		assert(failed);
 	});